+++ docker/Dockerfile.php
@@ -2,12 +2,22 @@ FROM php:7.4-fpm
 
 ENV COMPOSER_HOME /tmp
 
-RUN apt-get update && \
-    apt-get install -y git libpq-dev libzip-dev unzip libicu-dev locales && \
-    pecl install xdebug && \
-    docker-php-ext-configure intl && \
-    docker-php-ext-install -j$(nproc) intl gettext zip pdo pdo_pgsql && \
-    docker-php-ext-enable xdebug
+RUN apt-get update && apt-get install -y \
+        git \
+        libpq-dev \
+        libzip-dev \
+        unzip \
+        libicu-dev \
+        locales \
+        libfreetype6-dev \
+        libjpeg62-turbo-dev \
+        libpng-dev \
+        libwebp-dev \
+    && pecl install xdebug \
+    && docker-php-ext-configure intl \
+    && docker-php-ext-configure gd --with-webp --with-jpeg --with-freetype \
+    && docker-php-ext-install -j$(nproc) intl gettext zip pdo pdo_pgsql gd \
+    && docker-php-ext-enable xdebug
 
 RUN echo 'en_GB.UTF-8 UTF-8' >> /etc/locale.gen && \
     echo 'fr_FR.UTF-8 UTF-8' >> /etc/locale.gen && \