+++ src/models/dao/Link.php
@@ -218,4 +218,24 @@ class Link extends \Minz\DatabaseModel
         ]);
         return $statement->fetchAll();
     }
+
+    /**
+     * Return links with oldest fetched_at date.
+     *
+     * @param integer $number
+     *
+     * @return array
+     */
+    public function listByOldestFetching($number)
+    {
+        $sql = <<<SQL
+             SELECT * FROM links
+             ORDER BY fetched_at
+             LIMIT ?
+        SQL;
+
+        $statement = $this->prepare($sql);
+        $statement->execute([$number]);
+        return $statement->fetchAll();
+    }
 }